+++ test/clj_kondo/constant_condition_test.clj
@@ -109,11 +109,16 @@
      :col 16
      :level :warning
      :message "Condition always false"}]
-   (lint! "(let [a 4 b nil]
+   (lint! "(let [a 4 b (:k {})]
              (cond-> {}
                a (assoc :a a)
                b (assoc :b b)))"
-          config)))
+          config))
+  (testing "a nil literal is exempt, also as a cond-> test"
+    (is (empty? (lint! "(let [b nil]
+                          (cond-> {}
+                            b (assoc :b b)))"
+                       config)))))
 
 (deftest cond-test
   (assert-submaps2
@@ -388,21 +393,11 @@
     (is (empty? (lint! "(defn e [x] (when (or x nil) 1))" config)))))
 
 (deftest condition-always-false-test
-  (testing "nil literal in condition position"
-    (assert-submaps2
-     [{:file "<stdin>"
-       :row 1
-       :col 7
-       :level :warning
-       :message "Condition always false"}]
-     (lint! "(when nil 1)" config))
-    (assert-submaps2
-     [{:file "<stdin>"
-       :row 1
-       :col 5
-       :level :warning
-       :message "Condition always false"}]
-     (lint! "(if nil 1 2)" config)))
+  (testing "nil is exempt, an intentional dev toggle"
+    (is (empty? (lint! "(when nil 1)" config)))
+    (is (empty? (lint! "(if nil 1 2)" config)))
+    (is (empty? (lint! "(let [flag nil] (when flag 1))" config)))
+    (is (empty? (lint! "(when-some [x nil] x)" config))))
   (testing "a key lookup that is provably nil"
     (assert-submaps2
      [{:file "<stdin>"
@@ -500,7 +495,7 @@
   (testing "but only for always-true, the subset it used to cover"
     (assert-submaps2
      '({:row 1 :col 18 :message "Condition always false"})
-     (lint! "(defn f [] (when nil 1))"
+     (lint! "(defn f [] (when (:k {}) 1))"
             (assoc-in config [:linters :condition-always-true :level] :off)))
     (assert-submaps2
      '({:row 1 :col 26 :message "Unreachable code"})
@@ -522,5 +517,5 @@
                        config)))
     (assert-submaps2
      '({:row 1 :message "Condition always false"})
-     (lint! "#_{:clj-kondo/ignore [:condition-always-true]} (defn f [] (when nil 1))"
+     (lint! "#_{:clj-kondo/ignore [:condition-always-true]} (defn f [] (when (:k {}) 1))"
             (assoc-in config [:linters :redundant-ignore :level] :off)))))